--- application/src/main/java/com/github/pietergroenendijk/notifications/AppNotifier.java
@@ -1,8 +1,0 @@
-package com.github.pietergroenendijk.notifications;
-
-public class AppNotifier implements Notifier {
-    @Override
-    public void send(NotificationTask task) {
-
-    }
-}
