--- application/src/main/java/com/github/pietergroenendijk/notifications/EmailNotifier.java
@@ -1,8 +1,0 @@
-package com.github.pietergroenendijk.notifications;
-
-public class EmailNotifier implements Notifier {
-    @Override
-    public void send(NotificationTask task) {
-
-    }
-}
