--- application/src/main/java/com/github/pietergroenendijk/notifications/FineTresholdNotificationTaskData.java
@@ -1,5 +1,0 @@
-package com.github.pietergroenendijk.notifications;
-
-public class FineTresholdNotificationTaskData {
-
-}
