+++ application/src/main/java/com/github/pietergroenendijk/notifications/NotificationTask.java
@@ -1,5 +1,7 @@
 package com.github.pietergroenendijk.notifications;
 
+import com.github.pietergroenendijk.notifications.strategy.NotificationStrategy;
+
 import java.time.LocalTime;
 
 public record NotificationTask(
