--- application/src/main/java/com/github/pietergroenendijk/notifications/Notifier.java
@@ -1,5 +1,0 @@
-package com.github.pietergroenendijk.notifications;
-
-public interface Notifier {
-    void send(NotificationTask task);
-}
