+++ application/src/main/java/com/github/pietergroenendijk/notifications/generators/finetreshold/FineTresholdNotificationTaskData.java
@@ -1,0 +1,5 @@
+package com.github.pietergroenendijk.notifications.generators.finetreshold;
+
+public class FineTresholdNotificationTaskData {
+
+}
