+++ application/src/main/java/com/github/pietergroenendijk/notifications/notifiers/AppNotifier.java
@@ -1,0 +1,10 @@
+package com.github.pietergroenendijk.notifications.notifiers;
+
+import com.github.pietergroenendijk.notifications.NotificationTask;
+
+public class AppNotifier implements Notifier {
+    @Override
+    public void send(NotificationTask task) {
+
+    }
+}
