+++ application/src/main/java/com/github/pietergroenendijk/notifications/notifiers/EmailNotifier.java
@@ -1,0 +1,10 @@
+package com.github.pietergroenendijk.notifications.notifiers;
+
+import com.github.pietergroenendijk.notifications.NotificationTask;
+
+public class EmailNotifier implements Notifier {
+    @Override
+    public void send(NotificationTask task) {
+
+    }
+}
