+++ application/src/main/java/com/github/pietergroenendijk/notifications/notifiers/Notifier.java
@@ -1,0 +1,7 @@
+package com.github.pietergroenendijk.notifications.notifiers;
+
+import com.github.pietergroenendijk.notifications.NotificationTask;
+
+public interface Notifier {
+    void send(NotificationTask task);
+}
