+++ application/src/main/java/com/github/pietergroenendijk/notifications/strategy/NotificationStrategy.java
@@ -1,4 +1,7 @@
-package com.github.pietergroenendijk.notifications;
+package com.github.pietergroenendijk.notifications.strategy;
+
+import com.github.pietergroenendijk.notifications.NotificationTask;
+import com.github.pietergroenendijk.notifications.notifiers.Notifier;
 
 public class NotificationStrategy implements Notifier {
     private final Notifier[] channels;
