+++ application/src/main/java/com/github/pietergroenendijk/notifications/strategy/NotificationStrategyFactory.java
@@ -1,4 +1,9 @@
-package com.github.pietergroenendijk.notifications;
+package com.github.pietergroenendijk.notifications.strategy;
+
+import com.github.pietergroenendijk.notifications.notifiers.AppNotifier;
+import com.github.pietergroenendijk.notifications.notifiers.EmailNotifier;
+import com.github.pietergroenendijk.notifications.notifiers.Notifier;
+import com.github.pietergroenendijk.notifications.notifiers.SMSNotifier;
 
 public class NotificationStrategyFactory {
     private final EmailNotifier emailNotifier = new EmailNotifier();
